Summary: use list of types to display proper source type name.
